requirements.txt
numpy>=1.19.0
pillow>=8.0.0 